Fixed bug of difference of output path in reasoning and training mode
Output path must be a file in reasoning mode, while it must be a folder in training mode... Fixed.

There is still an open bug which causes powerloom to become unstable on second run without restarting application. It seems to be less frequent when there are 3 or more power loom files... Still need to fix this issue.